intergrations/Orion/requirements.txt
# Farm 5.0 Agent System - Dependencies

# Core Framework
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
aiofiles==23.2.1
pydantic==2.5.0

# AI & ML
openai==1.3.5
anthropic==0.7.7

# Email & Communication
google-auth==2.23.4
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.2.0
google-api-python-client==2.109.0
emails==0.6.0

# Database
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9
asyncpg==0.29.0

# Utilities
python-dotenv==1.0.0
requests==2.31.0
pandas==2.1.4
numpy==1.24.4
python-dateutil==2.8.2

# Security
cryptography==41.0.7
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Monitoring & Logging
prometheus-client==0.19.0
structlog==23.2.0

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2

# Development
black==23.11.0
flake8==6.1.0
mypy==1.7.1